feat(cli): resolve-display-name — the naming oracle outside Obsidian
Req: f17f7c57-d3b6-42d3-916e-8d59bc8447c5

`exocortex resolve-display-name <target>` prints the composed name an asset renders
as. Closes the UI/CLI Parity (#3417) gap measured 2026-08-15: a grep across the whole
--help tree (root + 8 subcommands) for displayname|display-name|printname|print-name
returned 0 — there was no way to ask, outside Obsidian, what an asset is called.

The gap became tangible when omitLabel shipped (v16.219.0): ems__Action assets are
created WITHOUT exo__Asset_label and their name is composed per-render by spec
f39ece43, so the only ways to check it were the human eye and the plugin's jest
harness — neither reachable from CI, the autonomous loop, or a machine without the
plugin installed.

## The engine moved to core behind a two-method port

Coupling was measured BEFORE the design, not assumed. Three of the four engine files
(1098 of 1871 lines) were already obsidian-free; the contact surface was exactly three
Obsidian calls — vault.getMarkdownFiles, metadataCache.getFileCache(...)?.frontmatter,
metadataCache.getFirstLinkpathDest. Crucially the scanner never uses a file object for
anything EXCEPT fetching frontmatter (collectDisplayNameSpec/collectDisplayNamePart take
frontmatter alone; a rule's sourceFile comes from the spec's own exo__Asset_uid, not a
path), so the port needs no file handle and collapses to two methods:

  listFrontmatter(): Iterable<Record<string, unknown>>
  resolveLinkpathFrontmatter(linkpath): Record<string, unknown> | null

Plugin supplies ObsidianVaultMetadataAdapter; CLI supplies FsVaultMetadataAdapter over
the existing FileSystemVaultAdapter — whose getAllFiles/getFrontmatter/getFirstLinkpathDest
already mirror the three Obsidian calls one-for-one. That correspondence is WHY a
two-method port was enough to reach parity without reimplementing anything.

The host-function registry keeps its (app, metadata) signature: core made the host
OPAQUE (it never inspects it) and forwards it, so the plugin types it as App on its side
and displayMatcherHostFunctions.ts is untouched.

## Zero caller changes — the shim IS the adapter boundary

31 files import the engine (14 src + 17 tests) — above the cascade cap. A plain
re-export would still have broken them, because tests construct
`new PrintNameRuleService(app)`. So the plugin-side shim is a SUBCLASS that takes App
and wraps it into the port. Result: not one of the 31 importers changed, not one of the
13 engine test-files changed, and the plugin's own construction site is byte-identical.

## Verification

- 10 display-name suites / 181 tests pass UNCHANGED (Scenario 4, empirically).
- Full plugin suite byte-identical to origin/main on this machine: 28 failed suites /
  313 passed / 3 failed tests / 5552 passed on BOTH branches (failures are `nanotar`
  env drift + an uninitialised submodule, in files this change never touches).
- Full CLI suite: the only failing suite is the pre-existing sparql-exo003, failing on
  pristine origin/main too.
- 6 new @req integration tests drive the REAL command over a REAL temp vault.
- Revert-verify, two axes, each restored and re-confirmed 6/6:
    * withhold the rule service from the resolver → 2 failed / 4 passed
    * make the port yield no frontmatter      → 2 failed / 4 passed
  Exactly the two spec-dependent scenarios go RED; the four negative controls
  (source=label, source=basename, missing target, outside-vault) stay GREEN.
- Live dogfood on an isolated full copy of vault-my: the composite created a
  label-less ems__Action and the oracle printed
  "⚡ ✅ Выпить Пантокальцин после обеда 2026-08-15 18:04" with source=spec and
  basename showing the bare UID that would leak without a spec.

## Scope

No display semantics changed — the move must be behaviourally neutral, so any output
divergence is a defect of the move rather than an improvement. The CLI registers no
host functions: their predicates are platform-specific and the engine is fail-closed,
so a spec naming an unregistered function simply never participates (the CLI
under-reports such a spec rather than guessing). Surfacing host functions to the CLI,
and any CI gate built on this verb, are deliberately separate steps.
