+++ packages/obsidian-plugin/src/domain/settings/ExocortexSettings.ts
@@ -1,53 +1,10 @@
-/**
- * Per-class display name template configuration
- */
-export interface DisplayNameSettings {
-  /** Global default template (used when no class-specific template exists) */
-  defaultTemplate: string;
-
-  /** Per-class template overrides (key = class name like "ems__Task") */
-  classTemplates: Record<string, string>;
-
-  /**
-   * TBox naming display-projection (RFC 78572fa9 Candidate B Phase 2). When ON (default),
-   * a TBox class/property definition displays as its COMPUTED symbolic projection
-   * `prefix#slug` (e.g. `ems#TaskPrototype`) — composing the namespace ontology's
-   * `exo__Ontology_shortName` with the entity's `exo__Slugable_slug` (with a label-derived
-   * fallback) — instead of its raw `exo__Asset_label`. Never stored; computed per-render.
-   * OFF restores the raw-label display (the revert-verify axis + a user escape hatch).
-   * Optional so a persisted settings object without it defaults ON (backward-compatible).
-   */
-  projectTBoxNames?: boolean;
-}
-
-/**
- * Default display name configuration
- *
- * FULL HOMOICONIC (#3838 part 3): classTemplates is EMPTY. Every per-class displayName
- * (the TaskPrototype/MeetingPrototype suffixes, the status/blocked prefixes, the DailyNote
- * basename) is now sourced ENTIRELY from vault exo__DisplayNameSpec assets read by
- * PrintNameRuleService (PMBOK project 47d57acf, req b4ee3caa) — no TS cold-start seed.
- *
- * Why the seeds were removable (orchestrator-verified via the live 8-surface Phase-6 smoke
- * on v16.183.0 + revert-verified tests): vault specs already WIN over any classTemplate
- * (compileDisplayNameSpecs registers them as HIGH-priority PrintNameRules), so the seeds were
- * dead-overridden once the vault scanned. Coverage lives in the vault — the unconditional
- * TaskPrototype spec e20fda38 and the MeetingPrototype spec d7bdca33 render the " (…Prototype)"
- * suffix; the pn__DailyNote seed was already DEAD (real DailyNotes key exo__Instance_class by
- * the bare UID [[b04e7a3e]], never the label pn__DailyNote — dn-2110 sweep). The only remaining
- * difference is a brief cold-start window before the vault scan — identical eventual-consistency
- * to the status emojis, which are also vault-served and already re-render on store-settle. No
- * steady-state user-visible change.
- *
- * `defaultTemplate` is the base fallback (a plain label), NOT a homoiconic seed — kept.
- */
-export const DEFAULT_DISPLAY_NAME_SETTINGS: DisplayNameSettings = {
-  defaultTemplate: "{{exo__Asset_label}}",
-
-  classTemplates: {},
-
-  projectTBoxNames: true,
-};
+// Moved to packages/core alongside the displayName engine (req f17f7c57). Re-exported from
+// the original path so the 9 importers of these symbols stay unchanged — the shim IS the
+// zero-regression mechanism, not a convenience.
+import type { DisplayNameSettings } from "@kitelev/exocortex-core";
+import { DEFAULT_DISPLAY_NAME_SETTINGS } from "@kitelev/exocortex-core";
+export type { DisplayNameSettings };
+export { DEFAULT_DISPLAY_NAME_SETTINGS };
 
 /**
  * Configuration for which channels a log level should route to.
